pyproject.toml
[project]
name = "gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "tiktoken>=0.11.0",
    "torch>=2.8.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
